// data-hooks/reports/compareBusStop/types.ts
export type CompareBusStopDataTypes = {
  date: string
  plantId: number
  employeeNo: string
  title: string
  firstName: string
  lastName: string
  periodOfDay: string
  registerBusLineName: string
  registerBusStopName: string
  registerDateTime: string
  busLineName: string
  busStopName: string
  time: string
}